### 2021-03-14 #3104 @whedon
[PRE REVIEW]: Traja: A Python toolbox for animal trajectory analysis
**Submitting author:** @justinshenk (<a href="http://orcid.org/0000-0002-0664-7337">Justin Shenk</a>)
**Repository:** <a href="https://github.com/traja-team/traja" target ="_blank">https://github.com/traja-team/traja</a>
**Version:** v0.2.5
