### 2021-03-14 #3104 @arfon
👋 @jni would you be willing to edit this submission for JOSS? I realized that #3103 might be a better fit for @cMadan.
