### 2021-03-15 #3104 @jni
Hi @justinshenk,

Thanks for your submission. While I look for reviewers (feel free to suggest any here, without @-mentioning them), could you please:

a) fix the references in the paper as suggested by [this comment](https://github.com/openjournals/joss-reviews/issues/3104#issuecomment-798969286) from our bot,
b) make sure you cite appropriate papers for any software you use. For example, I don't see citations for matplotlib, seaborn, or NumPy, all of which I *think* your software uses directly? And,
c) re-work the paper to move most of the content, which discusses the API in details, to the documentation? The paper should contain a succinct, high-level description of the software's functionality, not details of the API. For more, please read the ["What should my paper contain section](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) of our documentation.

Thank you!
