+++ paper/paper.bib
@@ -29,50 +29,6 @@
 }
 
 
-@article{socialways,
-  author    = {Javad Amirian and
-               Jean{-}Bernard Hayet and
-               Julien Pettr{\'{e}}},
-  title     = {Social Ways: Learning Multi-Modal Distributions of Pedestrian Trajectories
-               with GANs},
-  journal   = {CoRR},
-  volume    = {abs/1904.09507},
-  year      = {2019},
-  url       = {http://arxiv.org/abs/1904.09507},
-  archivePrefix = {arXiv},
-  eprint    = {1904.09507},
-  timestamp = {Fri, 26 Apr 2019 13:18:53 +0200},
-  biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1904-09507},
-  bibsource = {dblp computer science bibliography, https://dblp.org}
-}
-
-@article{next,
-  author    = {Junwei Liang and
-               Lu Jiang and
-               Juan Carlos Niebles and
-               Alexander G. Hauptmann and
-               Li Fei{-}Fei},
-  title     = {Peeking into the Future: Predicting Future Person Activities and Locations
-               in Videos},
-  journal   = {CoRR},
-  volume    = {abs/1902.03748},
-  year      = {2019},
-  url       = {http://arxiv.org/abs/1902.03748},
-  archivePrefix = {arXiv},
-  eprint    = {1902.03748},
-  timestamp = {Tue, 31 Dec 2019 11:53:05 +0100},
-  biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1902-03748},
-  bibsource = {dblp computer science bibliography, https://dblp.org}
-}
-
-@InProceedings{TraPHic,
-author = {Chandra, Rohan and Bhattacharya, Uttaran and Bera, Aniket and Manocha, Dinesh},
-title = {TraPHic: Trajectory Prediction in Dense and Heterogeneous Traffic Using Weighted Interactions},
-booktitle = {The IEEE Conference on Computer Vision and Pattern Recognition (CVPR)},
-month = {June},
-year = {2019}
-}
-
 @inproceedings{pandas,
   title={Data structures for statistical computing in python},
   author={McKinney, Wes and others},
@@ -80,6 +36,7 @@ year = {2019}
   volume={445},
   pages={51--56},
   year={2010},
+  doi ={10.25080/majora-92bf1922-00a},
   organization={Austin, TX}
 }
 
@@ -194,20 +151,6 @@ year = {2019}
 	title = {Hydroxytyrosol, the {Major} {Phenolic} {Compound} of {Olive} {Oil}, as an {Acute} {Therapeutic} {Strategy} after {Ischemic} {Stroke}. - {PubMed} - {NCBI}},
 	url = {https://www.ncbi.nlm.nih.gov/pubmed/31614692},
 	urldate = {2020-01-13}
-}
-
-@article{chandra_traphic:_2019,
-	title = {{TraPHic}: {Trajectory} {Prediction} in {Dense} and {Heterogeneous} {Traffic} {Using} {Weighted} {Interactions}},
-	shorttitle = {{TraPHic}},
-	url = {http://arxiv.org/abs/1812.04767},
[…]
