### 2021-03-16 traja-team/traja#77 @justinshenk
JOSS Paper Submission revision
We've been [requested ](https://github.com/openjournals/joss-reviews/issues/3104#issuecomment-799058851)to:

a) fix the references in the paper as suggested by this comment from our bot,
b) make sure you cite appropriate papers for any software you use. For example, I don't see citations for matplotlib, seaborn, or NumPy, all of which I think your software uses directly? And,
c) re-work the paper to move most of the content, which discusses the API in details, to the documentation? The paper should contain a succinct, high-level description of the software's functionality, not details of the API. For more, please read the [What should my paper contain](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) section of our documentation.

I fixed a (b60b5211609cc6d71e42aed07b347ef126f75e95) and b (a6c4816df68e3fb870209e81b850ad3b1ce09016). Can someone help with c? @Saran-nns @WolfByttner
