+++ paper/paper.md
@@ -12,7 +12,7 @@ authors:
   - name: Wolf Byttner
     affiliation: 3
     orcid: 0000-0002-9525-9730
-  - name: Saran Nambusubramaniyan
+  - name: Saranraj Nambusubramaniyan
     affiliation: 1
     orcid: 0000-0002-7314-0261
 affiliations:
@@ -51,7 +51,7 @@ A comprehensive source of documentation is provided on the home page
 
 The data used in this project includes animal trajectory data provided by [http://www.tecniplast.it](Tecniplast S.p.A.), manufacturer of laboratory animal equipment based in Varese, Italy, and Radboud University, Nijmegen, Netherlands. Tecniplast provided the mouse locomotion data collected with their Digital Ventilated Cages (DVC). The extracted coordinates of the mice requires further analysis with external tools. Due to lack of access to equipment, mouse home cage data is rather difficult to collect and analyze, thus few studies have been done onhomecage data. Furthermore, researchers who are interested in developing novel algorithms must implement from scratch much of the computational and algorithmic infrastructure for analysis and visualization. By packaging a library that is particularly useful for animal locomotion analysis, future researchers can benefit from access to a high-level interface and clearly documented methods for their work.
 
-[^1]: [http://www.tecniplast.it](http://www.tecniplast.it) 
+[^1]: [http://www.tecniplast.it](http://www.tecniplast.it)
 
 ## Design Principles
 
@@ -75,7 +75,7 @@ Traja depends on Matplotlib [@Hunter:2007] and Seaborn [@wasdom2020seaborn] for 
 
 The data samples presented here[^1] are in rectangular ($x$, $y$) Cartesian coordinates, reflecting the mouse home-cage (25x12.5 cm) dimensions. Analytical methods relevant to 2D rectilinear analysis of highly constrained spatial coordinates are thus primarily considered.
 
-![Traja package diagram](./images/diagram.jpg){width=80%} 
+![Traja package diagram](./images/diagram.jpg){width=80%}
 
 High volume data like animal trajectories has an increased tendency to be missing data due to data collection issues or noise. Filling in the missing data values, referred to as _data imputation_, is achieved with a wide variety of statistical or learning-based methods. As previously observed, data science projects typically require at least _95%_ of the time to be spent on cleaning, pre-processing and managing the data [@bosch_engineering_2021]. Therefore, several methods relevant to preprocessing animal data are demonstrated throuhghout the following sections.
 
@@ -142,10 +142,10 @@ __Table 1 `traja.trajectory` functionalities__
   to_shapely            |Convert to `Shapely` object
   transition_matrix     |Computes the transition matrix from binned trajectory
   transitions           | Computes the first-order Markov model for transitions between grid cells
-  
+
 ### The `plotting` module
 
-The `plotting` module contains all methods relevant to visualization of trajectories, features and models. A complete table of methods included as of writing are described in Table 
+The `plotting` module contains all methods relevant to visualization of trajectories, features and models. A complete table of methods included as of writing are described in Table
 
 __Table 2 `traja.plotting` functionalities__
 
@@ -180,7 +180,7 @@ The `rutils` module[^2] contains all methods relevant to interfacing R packages.
 -   trajr
 as well as respective plotting methods.
 
-[^2]: rutils available in version 0.2 - 0.2.3 and is removed in version 0.2.4 
+[^2]: rutils available in version 0.2 - 0.2.3 and is removed in version 0.2.4
 
 __Table 3 R packages with interfaces in Traja__
 
@@ -222,11 +222,11 @@ df.traja.rotate(angle=-20)
 ```
 and angle $\theta$ where $\theta \in R : \theta \in [-180,180]$.
 
-### Trip Grid 
+### Trip Grid
[…]
