### 2021-03-27 70a9291
[Paper] Remove API doc from paper
