+++ paper/paper.md
@@ -51,8 +51,6 @@ A comprehensive source of documentation is provided on the home page
 
 The data used in this project includes animal trajectory data provided by [http://www.tecniplast.it](Tecniplast S.p.A.), manufacturer of laboratory animal equipment based in Varese, Italy, and Radboud University, Nijmegen, Netherlands. Tecniplast provided the mouse locomotion data collected with their Digital Ventilated Cages (DVC). The extracted coordinates of the mice requires further analysis with external tools. Due to lack of access to equipment, mouse home cage data is rather difficult to collect and analyze, thus few studies have been done onhomecage data. Furthermore, researchers who are interested in developing novel algorithms must implement from scratch much of the computational and algorithmic infrastructure for analysis and visualization. By packaging a library that is particularly useful for animal locomotion analysis, future researchers can benefit from access to a high-level interface and clearly documented methods for their work.
 
-[^1]: [http://www.tecniplast.it](http://www.tecniplast.it)
-
 ## Design Principles
 
 Traja targets Python because of its popularity with data scientists.The library leverages the powerful pandas library, while adding methods specifically for trajectory analysis.
@@ -83,116 +81,12 @@ High volume data like animal trajectories has an increased tendency to be missin
 
 ## Overview of the Library
 
-The software is structured into several modules. This section surveys the structure of the codebase and elaborates on implementation strategies, but does not constitute a full documentation. For a detailed API reference, the reader is referred to the HTML documentation[^3]. The import classes are the `TrajaDataFrame` class located in `frame.py` and `TrajaAccessor` located in `accessor.py`. The root package diagram is shown in Figure 1.
+The software is structured into several modules. This section surveys the structure of the codebase and elaborates on implementation strategies, but does not constitute a full documentation. For a detailed API reference, the reader is referred to the HTML documentation hosted at  <https://traja.readthedocs.io>. The import classes are the `TrajaDataFrame` class located in `frame.py` and `TrajaAccessor` located in `accessor.py`. The root package diagram is shown in Figure 1.
 
-A high-level accessor (`TrajaAccessor`) provides access to data stored within the pandas dataframe, as well as methods within `trajectory` and `plotting` modules:
-
-```python
-import pandas as pd
-...
-@pd.api.extensions.register_dataframe_accessor("traja")
-class TrajaAccessor(object):
-    """Accessor for pandas DataFrame with trajectory-specific numerical and analytical functions.
-    Access with `df.traja`.
-    """
-    def __init__(self, pandas_obj):
-        self._validate(pandas_obj)
-        self._obj = pandas_obj
-    ...
-```
-
-Additionally `TrajaDataFrame` is a subclass of the pandas `DataFrame`, allowing instantiation of a `TrajaDataFrame` directly from an array of x,y coordinates:
-
-```python
-import numpy as np
-xy = np.array([[0,1,2],[1,2,3]])
-df = traja.TrajaDataFrame.from_xy(xy)
-```
-
+A high-level accessor (`TrajaAccessor`) provides access to data stored within the pandas dataframe.
 ### The `trajectory` module
 
 The `trajectory` module contains the methods relevant to preprocessing, analysis and modelling trajectories. A complete table of methods included as of writing are described in Table 1.
-
-__Table 1 `traja.trajectory` functionalities__
-
-|Name|Function|
-|----------------------|-------------------------------------------------|
-  angles                | Returns angles w.r.t. x-axis
-  calc_derivatives      |Computes the step displacement $\frac{ds}{dt}$ and displacement time
-  calc_displacement     |Computes the displacement $\frac{ds}{dt}$ between consecutive indices
[…]
