### 2021-03-27 25360ca
[Paper] Trim paper further, fix links
