+++ paper/paper.md
@@ -41,25 +41,12 @@ The data used in this project includes animal trajectory data provided by [http:
 
 ## Overview of the Library
 
-The software is structured into several modules. This section surveys the structure of the codebase and elaborates on implementation strategies, but does not constitute a full documentation. For a detailed API reference, the reader is referred to the HTML documentation hosted at  <https://traja.readthedocs.io>. The import classes are the `TrajaDataFrame` class located in `frame.py` and `TrajaAccessor` located in `accessor.py`.
+# Overview
+Traja targets Python because of its popularity with data scientists.The library leverages the powerful pandas library, while adding methods specifically for trajectory analysis.When importing traja, the traja namespace registers itself within the pandas dataframes namespace via df.traja.
 
-A high-level accessor (`TrajaAccessor`) provides access to data stored within the pandas dataframe.
-## Design Principles
+The software is structured into three parts. These provide functionality to transform, analyse and visualize trajectories. Full details are available at <https://traja.readthedocs.io/>.  The `trajectory` model provides analytical and preprocessing functionalities. The `models` subpackage provides both traditional and neural network-based tools to determine trajectory properties. The `plotting` module allows visualizing trajectories in various ways. 
 
-Traja targets Python because of its popularity with data scientists.The library leverages the powerful pandas library, while adding methods specifically for trajectory analysis.
-
-When importing `traja`, the traja namespace registers itself within the pandas dataframes namespace via `df.traja`.
-
-This lets traja work directly with panda's internal objects:
-
-```python
- >>> df = pd.DataFrame({'x':[0,1,2], 'y':[2,3,4]})
- >>> df.traja.center
- (1.0, 3.0)
- >>> df.traja.plot()
-```
-
-Data, e.g., x and y coordinates, are stored as one-dimensional labelled arrays as instances of the pandas native `Series` class. Further, subclassing the pandas `DataFrame` allows creating an API that mirrors the pandas API which is familiar to most data scientists, thus reducing the barrier for entry while providing methods and properties specific to trajectories for rapid prototyping.
+Data, e.g., x and y coordinates, are stored as one-dimensional labelled arrays as instances of the pandas native `Series` class. Further, subclassing the pandas `DataFrame` allows providing an API that mirrors the pandas API which is familiar to most data scientists, thus reducing the barrier for entry while providing methods and properties specific to trajectories for rapid prototyping.
 Traja depends on Matplotlib [@Hunter:2007] and Seaborn [@waskom2020seaborn] for plotting and NumPy [@harris2020array] for computation.
 
 ## Mouse Locomotion Data
@@ -77,7 +64,7 @@ $$T_k = \{P_{k1}, P_{k2},...\}$$
 
 where $P_{ki}(i\geq 1)$ is a point in the trajectory.
 
-Generating spatial trajectory data via a random walk is possible by sampling from a distribution of angles and step sizes [@kareiva_analyzing_1983,@mclean_trajr:_2018]. A correlated random walk (Figure [2](#fig:generated){reference-type="ref" reference="fig:generated"}) is generated with `traja.generate`.
+Generating spatial trajectory data via a random walk is possible by sampling from a distribution of angles and step sizes [@kareiva_analyzing_1983,@mclean_trajr:_2018]. A correlated random walk (Figure [3](#fig:generated){reference-type="ref" reference="fig:generated"}) is generated with `traja.generate`.
 
 ## Spatial Transformations
[…]
