### 2021-03-27 e63c156
[Paper] Fix image width, remove documentation
