+++ paper/paper.md
@@ -88,13 +88,13 @@ Allowing computation on discrete variables rather than continuous ones has sever
 Creation of an $M * N$ grid allows mapping trajectory $T_k$ onto uniform
 grid cells. Generalizing the nomenclature of [@wang_modeling_2017] to rectangular grids, $C_{mn}(1\leq{m}\leq M; 1\leq{n}\leq{N})$ denotes the cell in row $m$ and column $n$ of the grid. Each point $P_{ki}$ is assigned to a cell $C(m,n)$. The result is a two-dimensional image $M*N$ image $I_k$, where the value of pixel $I_k(m,n)(1\leq{m,n}\leq{M})$ indicates the relative number of points assigned to cell $C_{mn}$. Partionining of spatial position into separate grid cells is typically preceded by generation of hidden Markov models [@jeung_mining_2007] (see below).
 
-![Visualization of heat map from bins generated with `df.trip_grid`. Note regularly spaced artifacts (bright yellow) in this sample due to a bias in the sensor data interpolation. This type of noise can be minimized by thresholding or using a logarithmic scale (`traja.trip_grid(trj, log=True)`, as shown above.[]{label="fig:heatmap"}](./images/tripgrid.png){#fig:heatmap width=80%}
+![Visualization of heat map from bins generated with `df.trip_grid`. Note regularly spaced artifacts (bright yellow) in this sample due to a bias in the sensor data interpolation. This type of noise can be minimized by thresholding or using a logarithmic scale (`traja.trip_grid(trj, log=True)`, as shown above.[]{label="fig:heatmap"}](./images/tripgrid.png){#fig:heatmap width=50%}
 
 ### Feature Scaling
 
 Feature scaling is common practice for preprocessing data for machine learning [@grus_data_2015] and is essential for even application of methods to attributes. For example, a high dimensional feature vector $\mathbf{x} \in \mathbb{R}^n$ where some attributes are in $(0,100)$ and others are in $(-1,1)$ would lead to biases in the treatment of certain attributes. To limit the dynamic range for multiple data instances simultaneously, scaling is applied to a feature matrix $X = \{\mathbf{x_1}, \mathbf{x_2}, ..., \mathbf{x_N}\} \in \mathbb{R}^{n\times{N}}$, where $n$ is the number of instances.
 
-![Generation of a random walk[]{label="fig:generated"}](./images/generate.png){#fig:generated width="\textwidth"}
+![Generation of a random walk[]{label="fig:generated"}](./images/generate.png){#fig:generated width=80%}
 
 ### Min-Max Scaling
 
@@ -215,7 +215,7 @@ the model.[]{label="fig:kmeans"}](./images/kmeans_pca-fortasyn.png){#fig:kmeans
 width=80%}
 
 ### Hierarchical Agglomerative Clustering
-Clustering spatial trajectories has broad applications. For mice, hierarchical agglomerative clustering can be used to identify similarities between groups, for example periodic activity and location visit frequency. Clustering actograms is possible with df.traja.plot_cluster.
+Clustering spatial trajectories has broad applications. For mice, hierarchical agglomerative clustering can be used to identify similarities between groups, for example periodic activity and location visit frequency.
 
 ### Gaussian Processes
 Gaussian Processes is a non-parametric method which can be used to model spatial trajectories. This method is not currently implemented in Traja
@@ -241,7 +241,7 @@ Traja implements the rectangular spatial grid version of HMM with transitions.
 The probability of transition from each cell to another cell is stored as a probability within the transition matrix. This can further be plotted with `plot_transition_matrix` (Figure [9](#fig:transitionmatrix){reference-type="ref" reference="fig:transitionmatrix"}).
 
 ### Convex Hull
[…]
