+++ paper/paper.md
@@ -64,7 +64,9 @@ $$T_k = \{P_{k1}, P_{k2},...\}$$
 
 where $P_{ki}(i\geq 1)$ is a point in the trajectory.
 
-Generating spatial trajectory data via a random walk is possible by sampling from a distribution of angles and step sizes [@kareiva_analyzing_1983,@mclean_trajr:_2018]. A correlated random walk (Figure [3](#fig:generated){reference-type="ref" reference="fig:generated"}) is generated with `traja.generate`.
+Generating spatial trajectory data via a random walk is possible by sampling from a distribution of angles and step sizes [@kareiva_analyzing_1983,@mclean_trajr:_2018]. A correlated random walk (Figure [1](#fig:generated){reference-type="ref" reference="fig:generated"}) is generated with `traja.generate`.
+
+![Generation of a random walk[]{label="fig:generated"}](./images/generate.png){#fig:generated width=80%}
 
 ## Spatial Transformations
 Transformation of trajectories can be useful for comparing trajectories from various geospatial coordinates, data compression, or simply for visualization purposes.
@@ -76,7 +78,7 @@ $$\begin{bmatrix} x'\\y' \end{bmatrix} = \begin{bmatrix} cos\theta & i sin\theta
 with angle $\theta$ where $\theta \in R : \theta \in [-180,180]$.
 
 ### Trip Grid
-One strategy for compressing the representation of trajectories is binning the coordinates to produce an image as shown in Figure [1](#fig:tripgridalgo){reference-type="ref" reference="fig:tripgridalgo"}.
+One strategy for compressing the representation of trajectories is binning the coordinates to produce an image as shown in Figure [2](#fig:tripgridalgo){reference-type="ref" reference="fig:tripgridalgo"}.
 
 ![Trip grid image generation from mouse
 trajectory.](./images/trip_grid_algo.png){#fig:tripgridalgo width=80%}
@@ -93,8 +95,6 @@ grid cells. Generalizing the nomenclature of [@wang_modeling_2017] to rectangula
 ### Feature Scaling
 
 Feature scaling is common practice for preprocessing data for machine learning [@grus_data_2015] and is essential for even application of methods to attributes. For example, a high dimensional feature vector $\mathbf{x} \in \mathbb{R}^n$ where some attributes are in $(0,100)$ and others are in $(-1,1)$ would lead to biases in the treatment of certain attributes. To limit the dynamic range for multiple data instances simultaneously, scaling is applied to a feature matrix $X = \{\mathbf{x_1}, \mathbf{x_2}, ..., \mathbf{x_N}\} \in \mathbb{R}^{n\times{N}}$, where $n$ is the number of instances.
-
-![Generation of a random walk[]{label="fig:generated"}](./images/generate.png){#fig:generated width=80%}
 
 ### Min-Max Scaling
 
@@ -126,7 +126,7 @@ Smoothing can also be achieved with traja using Savitzky-Golay filtering with `s
 
 Trajectories can be resampled by time or by step length. This can be useful for aligning trajectories from various data sources and sampling rates or reducing the number of data points to improve computational efficiency. Care must be taken to select a time interval which maintains information on the significant behavior. If the minimal time interval observed is selected for the points, calculations will be computationally intractable for some systems. If too large of an interval is selected, we will fail to capture changes relevant to the target behavior in the data.
 
-Resampling by time is performed with `resample_time`. Rediscretizing by step length is performed with `rediscretize` (Figure [2](#fig:step){reference-type="ref" reference="fig:step"}).
[…]
