### 2021-03-29 #3104 @justinshenk
> Hi @justinshenk,
>
> Thanks for your submission. While I look for reviewers (feel free to suggest any here, without @-mentioning them), could you please:
>
> a) fix the references in the paper as suggested by [this comment](https://github.com/openjournals/joss-reviews/issues/3104#issuecomment-798969286) from our bot,
> b) make sure you cite appropriate papers for any software you use. For example, I don't see citations for matplotlib, seaborn, or NumPy, all of which I _think_ your software uses directly? And,
> c) re-work the paper to move most of the content, which discusses the API in details, to the documentation? The paper should contain a succinct, high-level description of the software's functionality, not details of the API. For more, please read the ["What should my paper contain section](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) of our documentation.
>
> Thank you!

Hi @jni,
Thanks for your helpful comments! The paper has been updated as requested, with API documentation removed and links fixed.

Looking forward to your further review.
