+++ paper/paper.bib
@@ -797,6 +797,7 @@ These data demonstrate that home cage monitoring is scalable and run in real tim
 	month = may,
 	year = {2019},
 	note = {arXiv: 1902.03748},
+	doi = {10.1109/cvprw.2019.00358},
 	keywords = {Computer Science - Computer Vision and Pattern Recognition}
 }
 
