+++ paper/paper.md
@@ -207,7 +207,7 @@ data.[]{label="fig:LDA"}](./images/lda_fortasyn-period.png){#fig:LDA
 width=80%}
 
 ### Clustering
-Clustering of trajectories is an extensive topic with applications in geospatial data, vehicle and pedestrian classification, as well as molecular identification. K-Means clustering is an iterative unsupervised learning method that assigns a label to data points based on a distance function [@bishop_pattern_2006].
+Clustering of trajectories is an extensive topic with applications in geospatial data, vehicle and pedestrian classification, as well as molecular identification. K-Means clustering is an iterative unsupervised learning method that assigns a label to data points based on a distance function [@bishop_pattern_2006] (Figure [10]).
 
 ![K-Means clustering on the results of the PCA shown above reveals a high accuracy
 of classification, with a few errors. Cluster labels are generated by
