### 2021-03-30 372778e
Added convex_hull property to traja dataframe, currently set to populate (as array) on generate(convex_hull=True), or by manually setting like df.convex_hull = df[['x', 'y']].  Updated documentation and added test.  Also added inside function to determine if point is in polygon, and function to determine colinearity of three points.
