+++ traja/frame.py
@@ -218,10 +218,10 @@ class TrajaCollection(TrajaDataFrame):
     ]
 
     def __init__(
-            self,
-            trjs: Union[TrajaDataFrame, pd.DataFrame, dict],
-            id_col: Optional[str] = None,
-            **kwargs,
+        self,
+        trjs: Union[TrajaDataFrame, pd.DataFrame, dict],
+        id_col: Optional[str] = None,
+        **kwargs,
     ):
         """Initialize with trajectories with x, y, and time columns.
 
@@ -270,9 +270,9 @@ class TrajaCollection(TrajaDataFrame):
     def plot(self, colors=None, **kwargs):
         """Plot collection of trajectories with colors assigned to each id.
 
-            >>> trjs = {ind: traja.generate(seed=ind) for ind in range(3)} # doctest: +SKIP
-            >>> coll = traja.TrajaCollection(trjs) # doctest: +SKIP
-            >>> coll.plot() # doctest: +SKIP
+        >>> trjs = {ind: traja.generate(seed=ind) for ind in range(3)} # doctest: +SKIP
+        >>> coll = traja.TrajaCollection(trjs) # doctest: +SKIP
+        >>> coll.plot() # doctest: +SKIP
 
         """
         return traja.plotting.plot_collection(
@@ -298,10 +298,10 @@ class TrajaCollection(TrajaDataFrame):
 
 class StaticObject(object):
     def __init__(
-            self,
-            x: Optional[float] = None,
-            y: Optional[float] = None,
-            bounding_box: Tuple[float] = None,
+        self,
+        x: Optional[float] = None,
+        y: Optional[float] = None,
+        bounding_box: Tuple[float] = None,
     ):
         ...
         pass
