### 2021-04-02 717680e
Corrected error/ommission from test_trajectory.py
