### 2021-04-03 c1b074d
Reduce test duration
