### 2021-04-21 #3202 @jni
@a-paxton @abigailmcgovern thank you for agreeing to review! You should have been invited to join this repo — check your emails or go to https://github.com/openjournals/joss-reviews/invitations to see them. This is just so that you can tick the tick-boxes on your checklist at the top of this issue. You can then follow our bot Whedon's advice above to unsubscribe from the repo.

As a reminder, here are JOSS's review guidelines:

https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Essentially, you make your way through the checklist above. If you have any concerns about any of the items on the list, raise them as a comment on this issue or as a new issue in the software repo, whichever you feel is most appropriate.

If you have any questions, please let me know! Again, you can simply comment on this issue, or email me if you prefer.
