### 2021-05-17 #3202 @AbigailMcGovern
## Functionality
> * [ ]  **Functionality:** Have the functional claims of the software been confirmed?

Both the paper and the documentation claim that the package is capable of trajectory prediction with RNNs (particuarly using LSTM). When I tried to execute the example in the docs I ran into some issues (see https://github.com/traja-team/traja/issues/83). It seems that this documentation doesn't relate to the version available of PyPI (0.2.3), which is a little confusing for a new user. I installed the development version (0.2.5), which was unideal owing to the requirement for a number of dependencies most of which (but not all - i.e., pip and scikit-learn) were in the environment file. Even once I downloaded this version, I ran into some issues.

I also wasn't able to find any reference to some of the analysis methods mentioned in the paper (autocorrelation, power spectrum, PCA, LDA) in either the documentation or the code base. Perhaps I missed something here? If so, It would be good if the documentation was made clearer in regards to this.

## Software paper
> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

As I understand it, there are no other packages with this functionality so I understand the omission. However, there are other packages and software available for producing the tracks (e.g., https://github.com/DeepLabCut/DeepLabCut, https://github.com/vivekhsridhar/tracktor) and I believe that these have some limited analysis functionality. It would be great if you could outline *how* your package might be complementary to these and perform tasks that these cannot. Not only would this give the reader a clearer understanding of the relevant part of the software ecosystem, but it would make the importance of your package clearer.

> * [ ]  **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)?

The authors need to reference PyTorch: https://github.com/pytorch/pytorch/blob/master/CITATION
