### 2021-05-17 a4ccd0f
simplify assertions
