+++ requirements/extra.txt
@@ -1,8 +1,18 @@
 # extended list of package dependencies to reach full functionality
 
-psutil
 plotly
 pytest
 h5py
 ipython
 pre-commit
+shapely
+scipy>=1.4.1
+scikit-learn
+fastdtw
+plotly
+networkx
+seaborn
+torch
+h5py
+numba>=0.50.0
+pyDOE2>=1.3.0
