### 2021-05-17 3130e20
Update docs, update dev requirements
