### 2021-05-18 1b8e1c1
Add plot methods
