+++ traja/plotting.py
@@ -477,7 +477,7 @@ def plot(
 
 
 def plot_periodogram(trj, coord: str = "y", fs: int = 1, interactive: bool = True):
-    """Plot power spectral density using a periodogram.
+    """Plot power spectral density of ``coord`` timeseries using a periodogram.
 
     Args:
         trj - Trajectory
@@ -487,6 +487,10 @@ def plot_periodogram(trj, coord: str = "y", fs: int = 1, interactive: bool = Tru
 
     Returns:
         Figure
+
+    .. note:: 
+
+        Convenience wrapper for :meth:`scipy.signal.periodogram`.
 
     """
     from scipy import signal
@@ -513,20 +517,21 @@ def plot_autocorrelation(
         trj - Trajectory
         coord - 'x' or 'y'
         unit - string, eg, 'Days'
-        sample_rate - sample rate
         xmax - max xaxis value
         interactive - Plot immediately
 
     Returns:
         Matplotlib Figure
 
-    """
-    from statsmodels import api as sm
+    .. doctest::
 
-    pos = trj[coord].values
-    acf = sm.tsa.acf(pos, nlags=len(pos))
-    lag = np.arange(len(pos)) / sample_rate
-    plt.plot(lag, acf)
+        >>> df.traja.plot_autocorrelation() #doctest: +SKIP
+
+    .. note::
+
+        Convenience wrapper for pandas :meth:`~pandas.plotting.autocorrelation_plot`.
+    """
+    pd.plotting.autocorrelation_plot(trj[coord])
     plt.xlim((0, xmax))
     plt.xlabel(f"Lags ({unit})")
     plt.ylabel("Autocorrelation")
@@ -534,6 +539,41 @@ def plot_autocorrelation(
         plt.show()
     return plt.gcf()
 
+
+def plot_pca(trj: TrajaDataFrame, id_col: str="id", bins: tuple = (8,8)):
+    """Plot PCA comparing animals ids by trip grids.
+    
+    Args:
+        trj - Trajectory
+        id_col - column representing animal IDs
+        bins - shape for binning trajectory into a trip grid
+
+    Returns:
+        fig - Figure
+    
+    """
+    from sklearn.decomposition import PCA
+
+    grids = []
+    for ID in trj[id_col].unique():
+        animal = trj[trj.ID==ID]
+        animal.drop(columns=[id_col],inplace=True)
+        grid = animal.traja.trip_grid(bins = bins, hist_only=True)[0]
+        grids.append(grid.flatten())
+
+    gridsarr = np.array(grids)
+    pca = PCA(n_components=2)
+    X_r = pca.fit(gridsarr).transform(gridsarr)
+
+    for idx, animal in enumerate(X_r):
+        plt.scatter(X_r[idx, 0], X_r[idx, 1], color=f'C{idx}', alpha=.8, lw=2, label=idx)
+
+    plt.title("PCA")
+    plt.legend(title=id_col, loc='best', shadow=False, scatterpoints=1)
+    plt.xlabel("Principal Component 1")
+    plt.ylabel("Principal Component 2")    
+
+    return plt.gcf()
 
 def plot_collection(
     trjs: Union[pd.DataFrame, TrajaDataFrame],
