### 2021-05-19 a0b240b
Update docs, paper, plotting
