+++ docs/environment.yml
@@ -4,7 +4,7 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - python=3.7
+  - python>=3.7
   - sphinx-gallery
   - sphinx
   - scikit-learn
