### 2021-05-19 e7f45b7
Update docs, paper, plotting
