### 2021-05-19 f3e122c
Update docs, paper, plotting
