+++ traja/plotting.py
@@ -42,8 +42,9 @@ __all__ = [
     "plot_contour",
     "plot_clustermap",
     "plot_flow",
-    "plot_quiver",
-    "plot_periodogram",
+    "plot_pca",
+    "plot_periodogram",    
+    "plot_quiver",  
     "plot_stream",
     "plot_surface",
     "plot_transition_graph",
@@ -540,20 +541,27 @@ def plot_autocorrelation(
     return plt.gcf()
 
 
-def plot_pca(trj: TrajaDataFrame, id_col: str="id", bins: tuple = (8,8)):
+def plot_pca(trj: TrajaDataFrame, id_col: str="id", bins: tuple = (8,8), three_dims: bool = False, ax = None):
     """Plot PCA comparing animals ids by trip grids.
     
     Args:
         trj - Trajectory
         id_col - column representing animal IDs
         bins - shape for binning trajectory into a trip grid
+        three_dims - 3D plot. Default: False (2D plot)
+        ax - Matplotlib axes (optional)
 
     Returns:
         fig - Figure
     
     """
     from sklearn.decomposition import PCA
+    from sklearn.preprocessing import StandardScaler
 
+
+    DIMS = 3 if three_dims else 2
+
+    # Bin trajectories to trip grids
     grids = []
     for ID in trj[id_col].unique():
         animal = trj[trj.ID==ID]
@@ -561,12 +569,27 @@ def plot_pca(trj: TrajaDataFrame, id_col: str="id", bins: tuple = (8,8)):
         grid = animal.traja.trip_grid(bins = bins, hist_only=True)[0]
         grids.append(grid.flatten())
 
+    # Standardize the data
     gridsarr = np.array(grids)
-    pca = PCA(n_components=2)
-    X_r = pca.fit(gridsarr).transform(gridsarr)
+    X = StandardScaler().fit_transform(gridsarr)
 
+    # PCA projection
+    pca = PCA(n_components=DIMS)
+    X_r = pca.fit(X).transform(X)
+
+    # Create plot axes
+    if DIMS == 3:
+        fig = plt.figure()
+        ax = fig.add_subplot(111, projection='3d')
+    if not ax:
+        _, ax = plt.subplots()
+    
+    # Visualize 2D projection
     for idx, animal in enumerate(X_r):
-        plt.scatter(X_r[idx, 0], X_r[idx, 1], color=f'C{idx}', alpha=.8, lw=2, label=idx)
+        if DIMS == 2:
+            ax.scatter(X_r[idx, 0], X_r[idx, 1], color=f'C{idx}', alpha=.8, lw=2, label=idx)
+        elif DIMS == 3:
+            ax.scatter(X_r[idx, 0], X_r[idx, 1], ax.scatter[idx,2], color=f'C{idx}', alpha=.8, lw=2, label=idx)
 
     plt.title("PCA")
     plt.legend(title=id_col, loc='best', shadow=False, scatterpoints=1)
