### 2021-05-19 a1a6b8d
Update docs, fix warning
