+++ docs/source/auto_examples/index.rst
@@ -1,36 +1,0 @@
-:orphan:
-
-Gallery
-==================
-
-A gallery of examples
-
-
-.. raw:: html
-
-    <div style='clear:both'></div>
-
-
-
-.. only :: html
-
- .. container:: sphx-glr-footer
-    :class: sphx-glr-footer-gallery
-
-
-  .. container:: sphx-glr-download
-
-    :download:`Download all examples in Python source code: auto_examples_python.zip </traja/docs/source/auto_examples/auto_examples_python.zip>`
-
-
-
-  .. container:: sphx-glr-download
-
-    :download:`Download all examples in Jupyter notebooks: auto_examples_jupyter.zip </traja/docs/source/auto_examples/auto_examples_jupyter.zip>`
-
-
-.. only:: html
-
- .. rst-class:: sphx-glr-signature
-
-    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
