+++ docs/source/generate.rst
@@ -3,7 +3,14 @@ Generate Random Walk
 
 Random walks can be generated using :func:`~traja.trajectory.generate`.
 
-.. autofunction:: traja.trajectory.generate
 
+.. ipython:: python :okwarning:
+
+    import traja
+    
+    # Generate random walk
+    df = traja.generate(1000)
+
+.. autofunction:: traja.trajectory.generate
 
 .. image:: https://raw.githubusercontent.com/justinshenk/traja/master/docs/source/_static/walk_screenshot.png
