+++ paper/paper.bib
@@ -630,6 +630,14 @@ The online version of this article (doi:10.1186/s40462-017-0097-x) contains supp
 	pages = {19--27}
 }
 
+@article{Mathisetal2018,
+    title={DeepLabCut: markerless pose estimation of user-defined body parts with deep learning},
+    author = {Alexander Mathis and Pranav Mamidanna and Kevin M. Cury and Taiga Abe  and Venkatesh N. Murthy and Mackenzie W. Mathis and Matthias Bethge},
+    journal={Nature Neuroscience},
+    year={2018},
+    url={https://www.nature.com/articles/s41593-018-0209-y}}
+
+
 @incollection{pytorch,
 title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
 author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
