+++ paper/paper.md
@@ -165,8 +165,7 @@ $$LI = \frac{RT}{LT + RT} $$
 
 where RT is the number of right turns observed and LT is the number of left turns observed. Turns are counted within a left turn angle $\in$ ($\theta$, 90) and right turn angle $\in(-\theta,-90)$. A turn is considered to have a minimal step length.
 
-## Advanced Techniques
-### Periodic Analysis
+## Periodic Analysis
 Periodic behaviors are a consequence of the circadian rhythm as well as observing expression of underlying cognitive traits. Some basic implementations of periodic analysis of mouse cage data are presented.
 
 ### Autocorrelation
@@ -201,9 +200,9 @@ PCA.[]{label="fig:3dpca"}](./images/pca_fortasyn-period-3d.png){#fig:3dpca
 width=80%}
 
 ### Clustering
-Clustering of trajectories is an extensive topic with applications in geospatial data, vehicle and pedestrian classification, as well as molecular identification. K-Means clustering is an iterative unsupervised learning method that assigns a label to data points based on a distance function [@bishop_pattern_2006] (Figure [9]).
+Clustering of trajectories is an extensive topic with applications in geospatial data, vehicle and pedestrian classification, as well as molecular identification. K-means clustering is an iterative unsupervised learning method that assigns a label to data points based on a distance function [@bishop_pattern_2006] (Figure [9]).
 
-![K-Means clustering on the results of the PCA shown above reveals a high accuracy
+![K-means clustering on the results of the PCA shown above reveals a high accuracy
 of classification, with a few errors. Cluster labels are generated by
 the model.[]{label="fig:kmeans"}](./images/kmeans_pca-fortasyn.png){#fig:kmeans
 width=80%}
@@ -235,7 +234,7 @@ Traja implements the rectangular spatial grid version of HMM with transitions.
 The probability of transition from each cell to another cell is stored as a probability within the transition matrix. This can further be plotted with `plot_transition_matrix` (Figure [10](#fig:transitionmatrix){reference-type="ref" reference="fig:transitionmatrix"}).
 
 ### Convex Hull
-The convex hull of a subtrajectory is the set $X$ of points in the Euclidean plane that is the smallest convex set to include X. For computational efficiency, a geometric k-simplex can be plotted covering the convex hull by converting to a Shapely object and using Shapely’s `convex_hull` method.
+The convex hull of a subtrajectory is the set $X$ of points in the Euclidean plane that is the smallest convex set to include $X$. For computational efficiency, a geometric k-simplex can be plotted covering the convex hull by converting to a Shapely object and using Shapely’s `convex_hull` method.
 
 ### Recurrent Neural Networks
 In recent years, deep learning has transformed the field of machine learning. For example, the current state of the art models for a wide range of tasks, including computer vision, speech to text, and pedestrian trajectory prediction, are achieved with deep neural networks. Neural networks are essentially sequences of matrix operations and elementwise function application based on a collection of computing units known as nodes or neurons. These units perform operations, such as matrix multiplication on input features of a dataset, followed by backpropagation of errors, to identify parameters useful for approximating a function.
