+++ paper/paper.bib
@@ -630,12 +630,40 @@ The online version of this article (doi:10.1186/s40462-017-0097-x) contains supp
 	pages = {19--27}
 }
 
+@software{vivek_hari_sridhar_2017_1134016,
+  author       = {Vivek Hari Sridhar},
+  title        = {vivekhsridhar/tracktor: Tracktor},
+  month        = dec,
+  year         = 2017,
+  publisher    = {Zenodo},
+  version      = {tracktor},
+  doi          = {10.5281/zenodo.1134016},
+  url          = {https://doi.org/10.5281/zenodo.1134016}
+}
+
+
+@article{wijeyakulasuriya_machine_2020,
+	title = {Machine learning for modeling animal movement},
+	volume = {15},
+	url = {https://doi.org/10.1371/journal.pone.0235750},
+	doi = {10.1371/journal.pone.0235750},
+	abstract = {Animal movement drives important ecological processes such as migration and the spread of infectious disease. Current approaches to modeling animal tracking data focus on parametric models used to understand environmental effects on movement behavior and to fill in missing tracking data. Machine Learning and Deep learning algorithms are powerful and flexible predictive modeling tools but have rarely been applied to animal movement data. In this study we present a general framework for predicting animal movement that is a combination of two steps: first predicting movement behavioral states and second predicting the animal’s velocity. We specify this framework at the individual level as well as for collective movement. We use Random Forests, Neural and Recurrent Neural Networks to compare performance predicting one step ahead as well as long range simulations. We compare results against a custom constructed Stochastic Differential Equation (SDE) model. We apply this approach to high resolution ant movement data. We found that the individual level Machine Learning and Deep Learning methods outperformed the SDE model for one step ahead prediction. The SDE model did comparatively better at simulating long range movement behaviour. Of the Machine Learning and Deep Learning models the Long Short Term Memory (LSTM) individual level model did best at long range simulations. We also applied the Random Forest and LSTM individual level models to model gull migratory movement to demonstrate the generalizability of this framework. Machine Learning and deep learning models are easier to specify compared to traditional parametric movement models which can have restrictive assumptions. However, machine learning and deep learning models are less interpretable than parametric movement models. The type of model used should be determined by the goal of the study, if the goal is prediction, our study provides evidence that machine learning and deep learning models could be useful tools.},
+	number = {7},
+	journal = {PLOS ONE},
+	author = {Wijeyakulasuriya, Dhanushi A. and Eisenhauer, Elizabeth W. and Shaby, Benjamin A. and Hanks, Ephraim M.},
+	year = {2020},
+	note = {Publisher: Public Library of Science},
+	pages = {1--30},
+}
+
+
 @article{Mathisetal2018,
     title={DeepLabCut: markerless pose estimation of user-defined body parts with deep learning},
     author = {Alexander Mathis and Pranav Mamidanna and Kevin M. Cury and Taiga Abe  and Venkatesh N. Murthy and Mackenzie W. Mathis and Matthias Bethge},
     journal={Nature Neuroscience},
     year={2018},
-    url={https://www.nature.com/articles/s41593-018-0209-y}}
+    url={https://www.nature.com/articles/s41593-018-0209-y}
+}
 
 
 @incollection{pytorch,
