+++ traja/plotting.py
@@ -507,8 +507,10 @@ def plot_periodogram(trj, coord: str = "y", fs: int = 1, interactive: bool = Tru
     vals = trj[coord].values
     f, Pxx = signal.periodogram(vals, fs=fs, window="hanning", scaling="spectrum")
     plt.title("Power Spectrum")
+    plt.plot(f, Pxx)
     if interactive:
-        plt.plot(f, Pxx)
+        plt.show()
+
     return plt.gcf()
 
 
@@ -516,7 +518,6 @@ def plot_autocorrelation(
     trj: TrajaDataFrame,
     coord: str = "y",
     unit: str = "Days",
-    sample_rate: float = 3.0,
     xmax: int = 1000,
     interactive: bool = True,
 ):
@@ -670,7 +671,7 @@ def plot_collection(
         colors = color_lookup
         labels = ids
 
-    fig, ax = plt.subplots()
+    _, ax = plt.subplots()
     lines = []
     for idx, id in enumerate(ids):
         trj = trjs[trjs[id_col] == id]
@@ -746,16 +747,21 @@ def plot_contour(
     contourplot_kws: dict = {},
     contourfplot_kws: dict = {},
     quiverplot_kws: dict = {},
+    ax: Axes = None,
     **kwargs,
 ) -> Axes:
     """Plot average flow from each grid cell to neighbor.
 
     Args:
+        trj: Traja DataFrame
         bins (int or tuple): Tuple of x,y bin counts; if `bins` is int, bin count of x,
                                 with y inferred from aspect ratio
+        filled (bool): Contours filled
+        quiver (bool): Quiver plot
         contourplot_kws: Additional keyword arguments for :meth:`~matplotlib.axes.Axes.contour`
         contourfplot_kws: Additional keyword arguments for :meth:`~matplotlib.axes.Axes.contourf`
         quiverplot_kws: Additional keyword arguments for :meth:`~matplotlib.axes.Axes.quiver`
+        ax (optional): Matplotlib Axes
 
     Returns:
         ax (:class:`~matplotlib.axes.Axes`): Axes of quiver plot
@@ -766,7 +772,8 @@ def plot_contour(
     X, Y, U, V = coords_to_flow(trj, bins)
     Z = np.sqrt(U * U + V * V)
 
-    fig, ax = plt.subplots()
+    if not ax:        
+        _, ax = plt.subplots()
 
     if filled:
         cfp = plt.contourf(X, Y, Z, **contourfplot_kws)
@@ -810,7 +817,7 @@ def plot_surface(
     fig = plt.figure()
     ax = fig.gca(projection="3d")
     ax.plot_surface(
-        X, Y, Z, cmap=matplotlib.cm.coolwarm, linewidth=0, **surfaceplot_kws
+        X, Y, Z, cmap= cmap, linewidth=0, **surfaceplot_kws
     )
 
     ax = _label_axes(trj, ax)
