### 2021-05-21 #3202 @justinshenk
Thank you @AbigailMcGovern for your review and valuable feedback!

The majority of edits are visible in https://github.com/traja-team/traja/pull/86.

The requested edits have been made in version 0.2.7, now available in PyPI for pip installation, including:

- updated environment.yml to include pip and scikit-learn
- added documentation for [autocorrelation](https://traja.readthedocs.io/en/latest/gallery/plot_autocorrelation.html?highlight=autocorrelation), power spectrum and [PCA](https://traja.readthedocs.io/en/latest/clustering.html#pca)
- removed LDA from the paper due to overlap with PCA
- reorganized the sections (eg, [Periodicity](https://traja.readthedocs.io/en/latest/periodicity.html)) of http://traja.readthedocs.org/
- fixed the documentation for the LSTM and fixed the error with the outputs
- fixed the headings in the paper to make the structure clearer for the reader
- added a reference in the paper for PyTorch
- other small improvements to the paper and documentation

I am very much looking forward to your further review.
