### 2021-05-21 #3202 @jni
Thanks for the update @justinshenk!

I think one additional item requested by @AbigailMcGovern is to add a bit more context about the existing state of the field. Looking at the new paper draft,

- there is a missing/misrendered reference, `vivek_hari_sridhar...` in the second paragraph,
- there's a few additional tools and corresponding papers that should probably be cited, listed in the [DLCUtils readme](https://github.com/DeepLabCut/DLCutils#clustering-tools-using-the-output-of-dlc). From a quick browse, [DLCAnalyzer](https://github.com/ETHZ-INS/DLCAnalyzer) (R), [B-SOID](https://github.com/YttriLab/B-SOiD) (Matlab) and [simba](https://github.com/sgoldenlab/simba) (Python) are the most relevant. You don't have to spend pages describing them, but a quick mention and overview of any distinguishing features should be included, in my opinion. For DLCAnalyzer and B-SOID it's obvious :wink:, for simba less so — but there are probably many design differences that you could highlight, e.g. Traja seems more targeted towards Pythonistas, while simba is a GUI application.

Other than that, this looks close to finished, thanks @AbigailMcGovern for the review and @justinshenk for iterating!

@a-paxton could we please get a review from you?
