+++ paper/paper.bib
@@ -630,6 +630,34 @@ The online version of this article (doi:10.1186/s40462-017-0097-x) contains supp
 	pages = {19--27}
 }
 
+@software{sgoldenlab_2021_4521178,
+  author       = {sgoldenlab and
+                  Jia Jie Choong and
+                  Simon Nilsson and
+                  Aasiya Islam and
+                  sophihwang26},
+  title        = {sgoldenlab/simba: SimBA: release v1.3},
+  month        = feb,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {v1.3},
+  doi          = {10.5281/zenodo.4521178},
+  url          = {https://doi.org/10.5281/zenodo.4521178}
+}
+
+@article {Hsu770271,
+	author = {Hsu, Alexander I. and Yttri, Eric A.},
+	title = {B-SOiD: An Open Source Unsupervised Algorithm for Discovery of Spontaneous Behaviors},
+	elocation-id = {770271},
+	year = {2020},
+	doi = {10.1101/770271},
+	publisher = {Cold Spring Harbor Laboratory},
+	abstract = {Capturing the performance of naturalistic behaviors remains a prohibitively difficult objective. Recent machine learning applications have enabled localization of limb position; however, position alone does not yield behavior. To provide a bridge from positions to actions and their kinematics, we developed Behavioral Segmentation of Open-field in DeepLabCut (B-SOiD). This unsupervised learning algorithm discovers natural patterns in position and extracts their inherent statistics. The cluster statistics are then used to train a machine learning algorithm to classify behaviors with greater speed and accuracy due to improved ability to generalize from subject to subject. Through the application of a novel fram-eshift paradigm, B-SOiD provides the fast temporal resolution required for comparison with neural activity. B-SOiD also provides high-resolution behavioral measures such as stride and grooming kinematics, that are difficult but critical to obtain, particularly in the study of pain, compulsion, and other neurological disorders. This open-source platform surpasses the current state of the field in its improved analytical accessibility, objectivity, and ease of use.},
+	URL = {https://www.biorxiv.org/content/early/2020/03/07/770271},
+	eprint = {https://www.biorxiv.org/content/early/2020/03/07/770271.full.pdf},
+	journal = {bioRxiv}
+}
+
 @software{vivek_hari_sridhar_2017_1134016,
   author       = {Vivek Hari Sridhar},
   title        = {vivekhsridhar/tracktor: Tracktor},
@@ -655,7 +683,6 @@ The online version of this article (doi:10.1186/s40462-017-0097-x) contains supp
 	note = {Publisher: Public Library of Science},
 	pages = {1--30},
 }
-
 
 @article{Mathisetal2018,
     title={DeepLabCut: markerless pose estimation of user-defined body parts with deep learning},
