### 2021-05-27 #3202 @justinshenk
Thanks for the helpful suggestions @jni!

The references and descriptions of other libraries relevant to Traja have been added in https://github.com/traja-team/traja/commit/1eaf5fb8ea3486ff3b08a2e7fb04b82526f8ecfa as requested.

Looking forward to the further review.
