### 2021-05-27 #3202 @a-paxton
## Functionality

> * [ ]  **Functionality:** Have the functional claims of the software been confirmed?

- I would strongly recommend including accessible default color palettes for plotting. (For example, the current default plotting options for the 3D surface plot is not colorbind-friendly; I would suggest something like viridis instead.)
- I submitted issues #87, #88, and #89 while testing the package from a clean conda env using the conda-forge installation instructions. After I also found `traja.contrib.rdp()` wasn't working, I checked the last upload date for traja on conda-forge (https://anaconda.org/conda-forge/traja) and realized that it was last uploaded over 9 months ago. I'm guessing a lot of my issues may stem from the continued development that your team has been accomplishing during that time! Since the paper says that it can be installed via conda-forge, I'd strongly encourage updating the package there. I'm going to hold off on my progress through the documentation until after I hear back from the authors about that decision, since I'm concerned that I might run into similar functionality discrepancies.

## Documentation

> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

- The motivation is clearly described. However, in the README.md and readthedocs Description, please change "handicapped" to more inclusive language (e.g., "restricted").

> * [ ]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

- The authors provide examples using toy data. Does that count here as "real-world analysis problems," or does that criterion mean that it needs to be real (not toy) data?
- To ensure computational reproducibility of the generated data (`df = traja.generate()`), I would recommend setting the seed on gallery examples.
- No example is given for `traja.plotting.animate()`, and I unfortunately wasn't able to get it to render. It sounds like a neat function, so it might be nice to include an example here.

> * [ ]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

- The documentation is great, but as a minor point, I would suggest replacing "w.r.t" with "with regard to" for maximal clarity.

> * [ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

- The guide to contributing is excellent, but I can't seem to find specific guidance for bugs/problems/issues or seeking support. (I understand if that the authors may simply be intending to use GitHub's systems for that; if so, including some guidance on that might be useful.)

## Software paper

> * [ ]  **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

- The summary is generally quite useful, but it's not clear from the Summary that Traja doesn't include extraction of trajectories from raw data (e.g., video, images). A brief explicit mention of that here would be helpful.

> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

- It would be nice to explicitly include a subsection (perhaps before the "Mouse Locomotion Data" section) that reviews ways that people can extract trajectories. I understand that this isn't part of the library, but it would be a valuable contribution for readers who may not be as familiar with the options available. It would also be a great opportunity to tie in some of the related software that @AbigailMcGovern and @jni mentioned (although some of the options that @jni mentioned may be better-suited to another comparison-focused section).

> * [ ]  **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

- The writing is generally good but should be double-checked for grammaticality (e.g., punctuation) and typos (e.g., missing spaces).
- Footnote 1 appears to include a typo or misplaced text ("## Spatial Trajectory...").

> * [ ]  **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)?

- I alluded to this above, but again, I had concerns about the reference list similar to that of @AbigailMcGovern and @jni. No need to reiterate them here---I completely understand that it might be in the works!
- The pandas library should be re-cited when mentioned again in the "Overview of the Library" section.
- Does the section in ll. 194-196 refer to previously published work or simply suggest directions for future work? If the former, please be sure to cite it.
- As someone who uses a lot of nonlinear time series analyses, I thought that the authors might be interested in adding some to the "Other Methods" section---like fractal analyses (e.g., Kerster, Rhodes, & Kello, 2016, Cognition; Huette, Kello, Rhodes, & Spivey, 2013, PLOS ONE) or recurrence quantification analysis (e.g., Ross, Warlaumont, Abney, Rigoli, & Balasubramaniam, 2016, Journal of Experimental Psychology: Human Perception and Performance; Shockley, Santana, & Fowler, 2003, Journal of Experimental Psychology: Human Perception and Performance). Of course, the works cited above are exclusively studies of human behavior, so the authors may wish to see whether there are related works in non-human fields (e.g., an analysis of bumble bee movements using recurrence quantification analysis in Ayers, Armsworth, & Brosi, 2015, Behavioral Ecology and Sociobiology). I don't think there's any need to include examples of their implementation, necessarily---just an opportunity to point readers toward potential interesting methods.
