+++ README.rst
@@ -60,7 +60,7 @@ videos.
 
 The goal of this package (and this document) is to aid biological
 researchers, who may not have extensive experience with Python, to
-analyze trajectories without being handicapped by a limited knowledge of
+analyze trajectories without being restricted by a limited knowledge of
 Python or programming. However, a basic understanding of Python is
 useful.
 
