### 2021-06-02 9a46232
Move colormap to viridis
