+++ docs/examples/plot_periodogram.py
@@ -8,5 +8,5 @@ Wrapper for pandas :meth:`scipy.signal.periodogram`.
 """
 import traja
 
-trj = traja.generate()
+trj = traja.generate(seed=0)
 trj.traja.plot_periodogram('x')
