+++ traja/plotting.py
@@ -1365,7 +1365,7 @@ def plot_transition_matrix(
     """
     if isinstance(data, np.ndarray):
         if data.shape[0] != data.shape[1]:
-            raise ValueException(
+            raise ValueError(
                 f"Ndarray input must be square transition matrix, shape is {data.shape}"
             )
         transition_matrix = data
@@ -1401,7 +1401,7 @@ def animate(trj: TrajaDataFrame, polar: bool = True, save: bool = False):
     fig = plt.figure(figsize=(8, 6))
     ax1 = plt.subplot(211)
 
-    fig.add_subplot(ax1)
+    fig.add_subplot(ax1)    
     if polar:
         ax2 = plt.subplot(212, polar="projection")
         ax2.set_theta_zero_location("N")
@@ -1429,54 +1429,56 @@ def animate(trj: TrajaDataFrame, polar: bool = True, save: bool = False):
     )
 
     def update(frame_number):
-        ind = frame_number % len(xy)
-        if ind < XY_STEPS:
-            scat.set_offsets(xy[:ind])
+        if frame_number < (XY_STEPS+2):
+            pass
         else:
-            prev_steps = max(ind - XY_STEPS, 0)
-            scat.set_offsets(xy[prev_steps:ind])
+            ind = frame_number % len(xy)
+            if ind < XY_STEPS:
+                scat.set_offsets(xy[:ind])
+            else:
+                prev_steps = max(ind - XY_STEPS, 0)
+                scat.set_offsets(xy[prev_steps:ind])
 
-        displacement_str = (
-            rf"$\bf{displacement[ind]:.2f}$"
-            if displacement[ind] >= DISPLACEMENT_THRESH
-            else f"{displacement[ind]:.2f}"
-        )
-
-        x, y = xy.iloc[ind]
-        ax1.set_title(
-            f"frame {ind} - distance (cm/0.25s): {displacement_str}\n"
-            f"x: {x:.2f}, y: {y:.2f}\n"
-            f"turn_angle: {turn_angle[ind]:.2f}"
-        )
-
-        if polar and ind > 1:
-            ax2.clear()
-            start_index = max(ind - POLAR_STEPS, 0)
-
-            theta = turn_angle[start_index:ind]
-            radii = displacement[start_index:ind]
-
-            hist, bin_edges = np.histogram(
-                theta, bins=np.arange(-180, 180 + bin_size, bin_size)
+            displacement_str = (
+                rf"$\bf{displacement[ind]:.2f}$"
+                if displacement[ind] >= DISPLACEMENT_THRESH
+                else f"{displacement[ind]:.2f}"
             )
-            centers = np.deg2rad(np.ediff1d(bin_edges) // 2 + bin_edges[:-1])
 
-            radians = np.deg2rad(theta)
+            x, y = xy.iloc[ind]
+            ax1.set_title(
+                f"frame {ind} - distance (cm/0.25s): {displacement_str}\n"
+                f"x: {x:.2f}, y: {y:.2f}\n"
+                f"turn_angle: {turn_angle[ind]:.2f}"
+            )
 
-            width = np.deg2rad(bin_size)
-            angle = radians if overlap else centers
-            height = radii if overlap else hist
-            max_height = displacement.max() if overlap else max(hist)
+            if polar and ind > 1:
+                ax2.clear()
+                start_index = max(ind - POLAR_STEPS, 0)
 
-            bars = ax2.bar(angle, height, width=width, bottom=0.0)
-            for idx, (h, bar) in enumerate(zip(height, bars)):
-                bar.set_facecolor(plt.cm.viridis(h / max_height))
-                bar.set_alpha(0.8 * (idx / POLAR_STEPS))
-            ax2.set_theta_zero_location("N")
[…]
