+++ docs/examples/plot_3d.py
@@ -9,7 +9,7 @@ import traja
 
 df = traja.TrajaDataFrame({"x": [0, 1, 2, 3, 4], "y": [1, 3, 2, 4, 5]})
 
-trj = traja.generate()
+trj = traja.generate(seed=0)
 ax = trj.traja.plot_3d(dist=15, labelpad=32, title="Traja 3D Plot")
 
 ########
@@ -19,3 +19,7 @@ ax = trj.traja.plot_3d(dist=15, labelpad=32, title="Traja 3D Plot")
 # `Matplotlib cmaps<https://matplotlib.org/examples/color/colormaps_reference.html>`_ are available
 
 trj.traja.plot_3d(dist=15, labelpad=32, title="Traja 3D Plot", cmap="jet")
+
+.. video:: https://github.com/traja-team/traja/blob/master/docs/_static/trajectory.mp4
+   :autoplay:
+   :nocontrols:
