### 2021-06-04 8198e83
[Docs] Update plotting and index
