+++ docs/source/conf.py
@@ -45,7 +45,7 @@ extensions = [
     "sphinx.ext.viewcode",
     "sphinx.ext.inheritance_diagram",
     "sphinx.ext.intersphinx",
-    "sphinx_gallery.gen_gallery",
+    "sphinx_gallery.gen_gallery"
 ]
 
 # continue doc build and only print warnings/errors in examples
@@ -75,6 +75,7 @@ sphinx_gallery_conf = {
     # "plot_gallery": None,
     "backreferences_dir": "reference",
     "within_subsection_order": FileNameSortKey,
+    "matplotlib_animations": True,
 }
 
 # Napoleon settings
