### 2021-06-04 0d312d4
[Docs] Update plotting and index
