+++ docs/examples/plot_animate.py
@@ -5,11 +5,12 @@ traja allows animating trajectories.
 """
 import traja
 
-df = traja.generate(100, seed=0)
+df = traja.generate(1000, seed=0)
 
 ###############################################################################
 # Plot a animation of trajectory
 # ==============================
 # An animation is generated using :func:`~traja.plotting.animate`.
 
-traja.plotting.animate(df, save=True) # saves to trajectory.mp4
+anim = traja.plotting.animate(df, save=True) # saves to trajectory.mp4
+anim
