+++ paper/paper.bib
@@ -40,18 +40,6 @@
   organization={Austin, TX}
 }
 
-  @Article{trajr,
-    title = {trajr: An R package for characterisation of animal
-      trajectories},
-    author = {Donald James McLean and Marta A. Skowron Volponi},
-    year = {2018},
-    journal = {Ethology},
-    volume = {124},
-    number = {6},
-    doi = {10.1111/eth.12739},
-    url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/eth.12739},
-  }
-
   @Article{adehabitat,
     title = {The package adehabitat for the R software: tool for the
       analysis of space and habitat use by animals},
@@ -924,6 +912,116 @@ These data demonstrate that home cage monitoring is scalable and run in real tim
 	keywords = {animal behaviour data, classification, clustering, dimensionality reduction, machine learning, predictive modelling, random forests, social networks, supervised learning, unsupervised learning},
 	pages = {203--220}
 }
+
+
+@article{kerster_spatial_2016,
+	title = {Spatial memory in foraging games},
+	volume = {148},
+	issn = {1873-7838},
+	doi = {10.1016/j.cognition.2015.12.015},
+	abstract = {Foraging and foraging-like processes are found in spatial navigation, memory, visual search, and many other search functions in human cognition and behavior. Foraging is commonly theorized using either random or correlated movements based on Lévy walks, or a series of decisions to remain or leave proximal areas known as "patches". Neither class of model makes use of spatial memory, but search performance may be enhanced when information about searched and unsearched locations is encoded. A video game was developed to test the role of human spatial memory in a canonical foraging task. Analyses of search trajectories from over 2000 human players yielded evidence that foraging movements were inherently clustered, and that clustering was facilitated by spatial memory cues and influenced by memory for spatial locations of targets found. A simple foraging model is presented in which spatial memory is used to integrate aspects of Lévy-based and patch-based foraging theories to perform a kind of area-restricted search, and thereby enhance performance as search unfolds. Using only two free parameters, the model accounts for a variety of findings that individually support competing theories, but together they argue for the integration of spatial memory into theories of foraging.},
+	language = {eng},
+	journal = {Cognition},
+	author = {Kerster, Bryan E. and Rhodes, Theo and Kello, Christopher T.},
+	month = mar,
+	year = {2016},
+	pmid = {26752603},
+	keywords = {Area-restricted search, Cognition, Feeding Behavior, Foraging, Humans, Search model, Spatial Behavior, Spatial memory, Spatial Memory, Spatial Navigation, Video Games},
+	pages = {85--96},
+}
+
+
+@article{ross_influence_2016,
+	title = {Influence of musical groove on postural sway},
+	volume = {42},
+	issn = {1939-1277(Electronic),0096-1523(Print)},
+	doi = {10.1037/xhp0000198},
[…]
