### 2021-06-04 9793179
Fix expected output for test
