### 2021-06-04 fb47508
[Paper] Add jaguar ref
