+++ paper/paper.md
@@ -51,7 +51,7 @@ Data, e.g., x and y coordinates, are stored as one-dimensional labelled arrays a
 Traja depends on Matplotlib [@Hunter:2007] and Seaborn [@waskom2020seaborn] for plotting and NumPy [@harris2020array] for computation.
 
 ### Trajectory Data Sources
-Trajectory data as time series can be extracted from a wide range of sources, including video processing tools as described above, GPS sensors for large animals or via homecage floor sensors, as described in the section below. The methods presented here are implemented for bivariate data (x and y coordinates) primarily to track animal centroids, however with some modification they could be extended to work in 3-dimensions and with body part locations as inputs. Traja is thus positioned at the end of the data scientists chain of tools with the hope of supporting prototyping novel data processing approaches.
+Trajectory data as time series can be extracted from a wide range of sources, including video processing tools as described above, GPS sensors for large animals or via homecage floor sensors, as described in the section below. The methods presented here are implemented for bivariate data (x and y coordinates) primarily to track animal centroids, however with some modification they could be extended to work in 3-dimensions and with body part locations as inputs. Traja is thus positioned at the end of the data scientists chain of tools with the hope of supporting prototyping novel data processing approaches. A sample dataset of jaguar movement [@morato_jaguar_2018] is provided in the `traja.dataset` subpackage.
 
 ## Mouse Locomotion Data
 The data samples presented here[^2] are in 2-dimensional location coordinates, reflecting the mouse home-cage (25x12.5 cm) dimensions. Analytical methods relevant to 2D rectilinear analysis of highly constrained spatial coordinates are thus primarily considered.
