+++ paper/paper.bib
@@ -959,6 +959,27 @@ journal = {Behavioral Ecology and Sociobiology},
 doi = {10.1007/s00265-015-1948-3}
 }
 
+@article{morato_jaguar_2018,
+	title = {Jaguar movement database: a {GPS}-based movement dataset of an apex predator in the {Neotropics}},
+	volume = {99},
+	copyright = {© 2018 The Authors. Ecology © 2018 The Ecological Society of America.},
+	issn = {1939-9170},
+	shorttitle = {Jaguar movement database},
+	url = {https://esajournals.onlinelibrary.wiley.com/doi/abs/10.1002/ecy.2379},
+	doi = {https://doi.org/10.1002/ecy.2379},
+	abstract = {The field of movement ecology has rapidly grown during the last decade, with important advancements in tracking devices and analytical tools that have provided unprecedented insights into where, when, and why species move across a landscape. Although there has been an increasing emphasis on making animal movement data publicly available, there has also been a conspicuous dearth in the availability of such data on large carnivores. Globally, large predators are of conservation concern. However, due to their secretive behavior and low densities, obtaining movement data on apex predators is expensive and logistically challenging. Consequently, the relatively small sample sizes typical of large carnivore movement studies may limit insights into the ecology and behavior of these elusive predators. The aim of this initiative is to make available to the conservation-scientific community a dataset of 134,690 locations of jaguars (Panthera onca) collected from 117 individuals (54 males and 63 females) tracked by GPS technology. Individual jaguars were monitored in five different range countries representing a large portion of the species’ distribution. This dataset may be used to answer a variety of ecological questions including but not limited to: improved models of connectivity from local to continental scales; the use of natural or human-modified landscapes by jaguars; movement behavior of jaguars in regions not represented in this dataset; intraspecific interactions; and predator-prey interactions. In making our dataset publicly available, we hope to motivate other research groups to do the same in the near future. Specifically, we aim to help inform a better understanding of jaguar movement ecology with applications towards effective decision making and maximizing long-term conservation efforts for this ecologically important species. There are no costs, copyright, or proprietary restrictions associated with this data set. When using this data set, please cite this article to recognize the effort involved in gathering and collating the data and the willingness of the authors to make it publicly available.},
+	language = {en},
+	number = {7},
+	urldate = {2021-06-04},
+	journal = {Ecology},
[…]
