### 2021-06-04 d2d6dae
[Docs] Fix animation
