### 2021-06-04 5a51f9b
[Docs] Fix animation
