### 2021-06-04 #3202 @justinshenk
Thank you @a-paxton for your in-depth review and helpful suggestions! I enjoyed reading the references you shared and plan to add support for several of these features in future versions of Traja.

Improvements as requested are provided below:

- Fixed conda-forge package in v0.2.8
- Default color palettes [changed](https://github.com/traja-team/traja/commit/9a462327d643b98512facb7c2c854d7df176f9d0) to viridis to be colorblind-friendly
- [Reference ](c46d803ac8d974bd14c0fb8fae9095b7b669dd8e) to a real-world dataset of jaguar coordinates is added to further highlight real-world relevance
- Seed has been [explicitily set](https://github.com/traja-team/traja/commit/4b8465fee6eabd5336b6ddc8b07276d5c02bb03a) on gallery examples
- The animate function now has an [example](https://traja.readthedocs.io/en/latest/gallery/animate.html) in the gallery
- "w.r.t." replaced with "with regards to" in https://github.com/traja-team/traja/commit/2a6e4ab44c889f777660d71db563933de5dbaa69
- Support/Bugs/Issues page [added](https://traja.readthedocs.io/en/latest/support.html) to the docs
- [Added explanation](https://github.com/traja-team/traja/commit/7890db542d661e7467ce4e8aa793876cfd47a18a#diff-e504eb580b095a7e65428b098183a581e475f0fb316db95287eacd7d4f344424R53) of trajectory data sources and extraction for improved clarity
- Fractal methods and other methods added; other improvements to the paper [are added](https://github.com/traja-team/traja/commit/7890db542d661e7467ce4e8aa793876cfd47a18a#)
- References to pandas and other libraries added

Thanks again for your review! If you see any other possible improvements please let me know.
