### 2021-06-04 0b8d920
Update README.rst
add conda-forge badge
