+++ README.rst
@@ -8,10 +8,13 @@ Traja |Python-ver| |Travis| |PyPI| |RTD| |Gitter| |Black| |License| |Binder| |Co
     :alt: Python 3.6+
 
 .. |Travis| image:: https://travis-ci.org/travis-team/traja.svg?branch=master
-    :target: https://travis-ci.org/travis-team/traja
+    :target: https://travis-ci.org/traja-team/traja
 
 .. |PyPI| image:: https://badge.fury.io/py/traja.svg
     :target: https://badge.fury.io/py/traja
+
+.. |Conda Version| image:: https://img.shields.io/conda/vn/conda-forge/traja.svg
+    :target: https://anaconda.org/conda-forge/traja
 
 .. |Gitter| image:: https://badges.gitter.im/traja-chat/community.svg
     :target: https://gitter.im/traja-chat/community
