### 2021-06-04 1349e7c
Update README.rst
Add conda-forge badge
