### 2021-06-04 traja-team/traja#83 @justinshenk
Closing as solved by https://github.com/traja-team/traja/pull/86
