+++ README.rst
@@ -73,7 +73,7 @@ If you use traja in your publications, please cite the repo
 
     @misc{justin_shenk_2019_3237827,
       author       = {Justin Shenk and
-                      Rüdiger Busche},
+                      the Traja development team},
       title        = {justinshenk/traja: v0.1.1},
       month        = jun,
       year         = 2019,
