### 2021-06-04 b70e6ea
Update README.rst
Revert to match zenodo for consistency
