### 2021-06-04 41d82de
Update README.rst
Make same as CITATION
