### 2021-06-13 81600aa
Started re-drafting pituitary gland with classes
