+++ traja/dataset/pituitary_gland.py
@@ -4,12 +4,13 @@ from numpy import exp
 from numba import jit
 from scipy.integrate import odeint
 from pyDOE2 import lhs
+import peakutils
 
 
 # PyTest will not compute coverage correctly for @jit-compiled code.
 # Thus we must explicitly suppress the coverage check.
 @jit
-def pituitary_ode(w, t, p):  # pragma: no cover
+def pituitary_ode_fletcher(w, t, p):  # pragma: no cover
     """
     Defines the differential equations for the pituirary gland system.
     To be used with scipy.integrate.odeint (this is the rhs equation).
@@ -70,10 +71,10 @@ def pituitary_ode(w, t, p):  # pragma: no cover
     return dv, dn, df, dc
 
 
-def compute_pituitary_gland_df_from_parameters(downsample_rate,
-                                               gcal, gsk, gk, gbk, gl, kc,
-                                               sample_id,
-                                               trim_start=20000):
+def compute_pituitary_gland_df_fletcher_from_parameters(downsample_rate,
+                                                        gcal, gsk, gk, gbk, gl, kc,
+                                                        sample_id,
+                                                        trim_start=20000):
     """
     Computes a Traja dataframe from the pituitary gland simulation.
 
@@ -114,7 +115,7 @@ def compute_pituitary_gland_df_from_parameters(downsample_rate,
 
     t = np.arange(0, 5000, 0.05)
     # print("Generating gcal={}, gsk={}, gk={}, gbk={}, gl={}, kc={}".format(gcal, gsk, gk, gbk, gl, kc))
-    wsol = odeint(pituitary_ode, w0, t, args=(p,), atol=abserr, rtol=relerr)
+    wsol = odeint(pituitary_ode_fletcher, w0, t, args=(p,), atol=abserr, rtol=relerr)
     df = pd.DataFrame(wsol, columns=['v', 'n', 'f', 'c'])
     df = df[trim_start:]
     df['ID'] = sample_id
@@ -142,10 +143,431 @@ def create_latin_hypercube_sampled_pituitary_df(downsample_rate=100, samples=100
     dataframes = []
     for sample_id, parameter in enumerate(parameters):
         gcal, gsk, gk, gbk, gl, kc = parameter
-        df = compute_pituitary_gland_df_from_parameters(downsample_rate,
-                                                        gcal, gsk, gk, gbk, gl, kc,
-                                                        sample_id)
+        df = compute_pituitary_gland_df_fletcher_from_parameters(downsample_rate,
+                                                                 gcal, gsk, gk, gbk, gl, kc,
+                                                                 sample_id)
         dataframes.append(df)
 
     num_samples = len(dataframes)
     return pd.concat(dataframes), num_samples
+
+
+def uniform_scale(val_min, val_mean, val_max):
+    return np.random.uniform(val_min / abs(val_mean), val_max / abs(val_mean))
+
+
+@jit
[…]
