+++ requirements/extra.txt
@@ -15,3 +15,4 @@ torch
 h5py
 numba>=0.50.0
 pyDOE2>=1.3.0
+peakutils>=1.3.3
