### 2021-06-23 #3202 @jni
Ok, so I'm happy that all reviewer comments have been addressed. It would be great if @AbigailMcGovern and @a-paxton could confirm this, but otherwise I'll go ahead and marked them as addressed this coming Monday.

In the meantime @justinshenk, I found one final typo in the paper:
- line 42: onhomecage data -> on home cage data
