### 2021-06-24 931c469
[Paper] Fix minor typos and formatting
