### 2021-06-24 5520db6
[Paper] Fix figure numbers and clarify text
