### 2021-06-24 4d48384
[Paper] Fix footnote numbering
