### 2021-06-24 f6893c7
[Paper] Fix footnote numbering
