### 2021-06-24 #3202 @justinshenk
Thank you @jni, for the final typo catch! Also, for the private recommendation regarding the author list. The following edits have been added to the paper:
- [Footnote number fixed](https://github.com/traja-team/traja/commit/f6893c7ac9fc9ba9f0532ed50c7018b0a10b42cb)
- [Figure numbers fixed and text clarified](https://github.com/traja-team/traja/commit/5520db63a0547c09c67f0f0796993bb8ffeb4f0a)
- [Minor typos and formatting fixed](https://github.com/traja-team/traja/commit/931c469a44e07ea8f83a820a19d8e7611ace574c)
- [Author list updated](https://github.com/traja-team/traja/commit/321003f38afaae95109ae6b80ad6c4fd6e91039f)

I believe it is ready for acceptance.
