### 2021-06-28 #3202 @a-paxton
My apologies for the delayed review on my end. I appreciate the extensive efforts from @justinshenk et al. to revise to the software, documentation, and manuscript in response to the previous round of reviews, and I'm happy to recommend the current version for acceptance!

PS: I noted a minor typo on line 232 ("hande-eye coordination" -> "hand-eye coordination"), but it's not important enough to hold up acceptance.
