### 2021-06-28 9a5e49f
[Paper] Fix minor typos
