### 2021-06-29 #3202 @jni
@justinshenk awesome, we are go! There is one final task, which is to update the version number for the current paper. Then I can set the release version and recommend acceptance! (The JOSS EICs actually do the acceptance — usually within a few hours of recommendation — and then publication is automatic and instant!)
