### 2021-06-29 #3202 @justinshenk
@jni I updated the package in Zendodo with v0.2.8. The new DOI is 10.5281/zenodo.5042781.
