### 2021-06-30 #3202 @jni
Sorry I wasn't clearer @justinshenk: Could you please tag a release with the latest changes (even though they are only to the paper) as v0.2.9? That way the source of the paper matches what is published in JOSS.
