### 2021-07-01 #3202 @justinshenk
Thanks for the clarification @jni. I tagged v0.2.9 (https://github.com/traja-team/traja/releases/tag/v0.2.9). Please let me know if anything else is missing.
