### 2021-07-01 #3202 @jni
@justinshenk hmm, looks like the new version didn't get picked up by Zenodo... Do you know why? Does it need to go out to PyPI? I always forget exactly how it works with Zenodo... Either way we need a DOI corresponding to this new version.
