### 2021-07-01 #3202 @justinshenk
I had been manually creating releases in Zenodo. Thanks to your suggestion, I found the easier way with GitHub integration. Now v0.2.9 should be picked up. But Zenodo is currently down :\. Hopefully you can see it by the time you look.
