+++ README.rst
@@ -7,7 +7,7 @@ Traja |Python-ver| |Travis| |PyPI| |Conda| |RTD| |Gitter| |Black| |License| |Bin
     :target: https://www.python.org/downloads/release/python-360/
     :alt: Python 3.6+
 
-.. |Travis| image:: https://travis-ci.org/travis-team/traja.svg?branch=master
+.. |Travis| image:: https://travis-ci.org/traja-team/traja.svg?branch=master
     :target: https://travis-ci.org/traja-team/traja
 
 .. |PyPI| image:: https://badge.fury.io/py/traja.svg
