+++ paper/paper.bib
@@ -677,7 +677,8 @@ The online version of this article (doi:10.1186/s40462-017-0097-x) contains supp
     author = {Alexander Mathis and Pranav Mamidanna and Kevin M. Cury and Taiga Abe  and Venkatesh N. Murthy and Mackenzie W. Mathis and Matthias Bethge},
     journal={Nature Neuroscience},
     year={2018},
-    url={https://www.nature.com/articles/s41593-018-0209-y}
+    url={https://www.nature.com/articles/s41593-018-0209-y},
+	doi = {10.1038/s41593-018-0209-y}
 }
 
 
