+++ requirements/extra.txt
@@ -2,7 +2,6 @@
 
 plotly
 pytest
-h5py
 ipython
 pre-commit
 shapely
@@ -12,6 +11,3 @@ fastdtw
 networkx
 seaborn
 torch
-h5py
-numba>=0.50.0
-pyDOE2>=1.3.0
