+++ requirements/docs.txt
@@ -16,5 +16,4 @@ pytest
 pytest-cov
 codecov
 readline
-h5py
 ipython
