### 2021-07-02 #3202 @justinshenk
New DOI for v0.2.10: 10.5281/zenodo.5061968
Hopefully good to go now :)
