### 2021-07-03 #3202 @arfon
@justinshenk – could you update the Seaborn reference to the JOSS paper please (https://doi.org/10.21105/joss.03021) ? (looking at their README I think this is the preferred citation now)
