+++ paper/paper.bib
@@ -1558,15 +1558,17 @@ CONCLUSION: The distinct characteristics of gait and gait variability in the MPT
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
 
-@software{waskom2020seaborn,
-    author       = {Michael Waskom and the seaborn development team},
-    title        = {mwaskom/seaborn},
-    month        = sep,
-    year         = 2020,
-    publisher    = {Zenodo},
-    version      = {latest},
-    doi          = {10.5281/zenodo.592845},
-    url          = {https://doi.org/10.5281/zenodo.592845},
+@article{Waskom2021,
+  doi = {10.21105/joss.03021},
+  url = {https://doi.org/10.21105/joss.03021},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {60},
+  pages = {3021},
+  author = {Michael L. Waskom},
+  title = {seaborn: statistical data visualization},
+  journal = {Journal of Open Source Software}
 }
 
 @Article{Hunter:2007,
