### 2021-07-03 15c3dd8
Bump version, fix citation
