### 2021-07-09 79845ef
Minor additions * Remove duplicate angle method "angles" * Add options to specify unit and lag in "calc_angle" * Check angle unit in calc_heading * Update tests, fix minor typos * Code formatting and fix flake8 issues
