### 2021-07-09 f560756
[Tests] Add github action
