+++ .github/workflows/tests.yaml
@@ -1,14 +1,11 @@
 name: Tests
 
 on:
-  release:
-    types: [published]
   push:
-    branches:
-      - [master]
+    branches: [master]
     tags: [v*]
   pull_request:
-    branches: [main, develop]
+    branches: [master]
 
 jobs:
   Linting: #see https://docs.github.com/en/actions/guides/building-and-testing-python
