### 2021-07-09 ff4ddf3
[Tests] Fix github action
