### 2021-07-09 4dd3e18
Fix variable reference, comparisons
