+++ traja/plotting.py
@@ -1,6 +1,7 @@
 import logging
 from collections import OrderedDict
 from datetime import timedelta
+import os
 from typing import Union, Optional, Tuple, List
 
 import matplotlib
