### 2021-07-09 0c02288
Fix convex_hull getter
