### 2021-07-09 e79a33e
Fix convex_hull getter
