### 2021-07-10 #3202 @Kevin-Mattheus-Moerman
@justinshenk I have proofread your paper and most looks good. Below are some final minor steps to work on:

- [ ] Please amend the title for the ZENODO archive version so that it matches the paper `Traja: A Python toolbox for animal trajectory analysis`, you will likely have to manually edit this.
- [ ] The text in figure 12 is very difficult to read. Would it be possible to increase the size of the fonts there? You can perhaps also increase the size of the figure itself. Figure 2 is borderline too in terms of font size, so consider update that one too.
