### 2021-07-12 a7c6d78
[Paper] Fix image size
