+++ requirements/docs.txt
@@ -14,6 +14,4 @@ torch
 pytest
 pytest-cov
 codecov
-readline
-h5py
 ipython
