### 2021-07-12 a2decf4
[Tests] Fix lib install
