+++ .github/workflows/tests.yaml
@@ -30,7 +30,7 @@ jobs:
       - name: Lint
         shell: bash -l {0}
         run: |
-            conda install flake8            
+            conda install flake8
             python -m flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
             python -m flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
       - name: Run pytest
@@ -39,10 +39,11 @@ jobs:
             pip install -r requirements/dev.txt
             pip install -r requirements/docs.txt
             pip install --upgrade pytest flake8 sphinx
+            pip install scipy --force-reinstall # for https://github.com/conda/conda/issues/6396
             pip install .
             conda install pytest
-            py.test . --cov-report=xml --cov=traja                    
-      - name: Upload coverage to Codecov 
+            py.test . --cov-report=xml --cov=traja
+      - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v1
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
