### 2021-07-17 d40f8d4
[Paper] Increase image width
